new journey class created
